[Soft-Float] - Initial Interpreter Implementation of Ps2's floating point unit specification.
This Pull Request implements the first take ever on real Soft-Float support in PCSX2.

This work is a combination or several efforts and researches done prior.

Credits:

- https://www.gregorygaines.com/blog/emulating-ps2-floating-point-nums-ieee-754-diffs-part-1/

- https://github.com/GitHubProUser67/MultiServer3/blob/main/BackendServices/CastleLibrary/EmotionEngine.Emulator/Ps2Float.cs

- https://github.com/Goatman13/pcsx2/tree/accurate_int_add_sub

- PCSX2 Team for their help and support in this massive journey.

This pull request should be tested with every games requiring a clamping/rounding mode (cf: GameDatabase).

Currently, this PR fixes on the interpreters:

- https://github.com/PCSX2/pcsx2/issues/354

- https://github.com/PCSX2/pcsx2/issues/11507

- https://github.com/PCSX2/pcsx2/issues/10519

- https://github.com/PCSX2/pcsx2/issues/8068

- https://github.com/PCSX2/pcsx2/issues/7642

- https://github.com/PCSX2/pcsx2/issues/5257

This is important to note, that this implementation, while technically fixing Gran Turismo 4 and Klonoa 2, makes the games crash due to very high floats being passed in the emu code, and failing at some points later in the process. This has not yet been ironed-out.

Other than that, this sets the floor for Soft-Float in PCSX2, a long awaited contribution.

+++ pcsx2/VUflags.h
@@ -4,10 +4,10 @@
 #pragma once
 #include "VU.h"
 
-extern u32  VU_MACx_UPDATE(VURegs * VU, float x);
-extern u32  VU_MACy_UPDATE(VURegs * VU, float y);
-extern u32  VU_MACz_UPDATE(VURegs * VU, float z);
-extern u32  VU_MACw_UPDATE(VURegs * VU, float w);
+extern u32  VU_MACx_UPDATE(VURegs * VU, uint32_t x);
+extern u32 VU_MACy_UPDATE(VURegs* VU, uint32_t y);
+extern u32 VU_MACz_UPDATE(VURegs* VU, uint32_t z);
+extern u32 VU_MACw_UPDATE(VURegs* VU, uint32_t w);
 extern void VU_MACx_CLEAR(VURegs * VU);
 extern void VU_MACy_CLEAR(VURegs * VU);
 extern void VU_MACz_CLEAR(VURegs * VU);
